removed some limits

+++ src/thread/default_attr.c
@@ -1,4 +1,4 @@
 #include "pthread_impl.h"
 
-unsigned __default_stacksize = DEFAULT_STACK_SIZE;
-unsigned __default_guardsize = DEFAULT_GUARD_SIZE;
+size_t __default_stacksize = DEFAULT_STACK_SIZE;
+size_t __default_guardsize = DEFAULT_GUARD_SIZE;
